Renaming chap-new.* to chap.*
- chap-md5.h is an internal headerfile, don't install
- At this point, there is nothing new with chap-new, it's the chap implementation
- chap_ms.c contains the specific Microsoft specific defines
