+++ pppd/plugins/radius/radius.c
@@ -38,7 +38,7 @@ static char const RCSID[] =
 
 #include <pppd/pppd.h>
 #include <pppd/options.h>
-#include <pppd/chap-new.h>
+#include <pppd/chap.h>
 #include <pppd/upap.h>
 #ifdef PPP_WITH_CHAPMS
 #include <pppd/chap_ms.h>
